requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
tensorflow
pillow
boto3
sqlalchemy
pymysql